+++ CSGL.Vulkan/Commands.cs
@@ -158,7 +158,7 @@ namespace CSGL.Vulkan {
                 Type t = typeof(DeviceCommands);
                 FieldInfo[] fields = t.GetFields();
                 for (int i = 0; i < fields.Length; i++) {
-                    string command = Vulkan.GetCommand(fields[i].FieldType);
+                    string command = Unmanaged.VK.GetCommand(fields[i].FieldType);
                     IntPtr ptr = device.GetProcAdddress(command);
                     if (ptr == IntPtr.Zero) continue;   //null if extension is not enabled
                     fields[i].SetValue(this, Marshal.GetDelegateForFunctionPointer(ptr, fields[i].FieldType));
@@ -199,7 +199,7 @@ namespace CSGL.Vulkan {
                 Type t = typeof(InstanceCommands);
                 FieldInfo[] fields = t.GetFields();
                 for (int i = 0; i < fields.Length; i++) {
-                    string command = Vulkan.GetCommand(fields[i].FieldType);
+                    string command = Unmanaged.VK.GetCommand(fields[i].FieldType);
                     IntPtr ptr = instance.GetProcAddress(command);
                     if (ptr == IntPtr.Zero) continue;   //null if extension is not enabled
                     fields[i].SetValue(this, Marshal.GetDelegateForFunctionPointer(ptr, fields[i].FieldType));